Add JavaResolver in SwiftJavaToolLib to resolve for ResolveCommand
